MI_requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
Deprecated==1.2.18
flatbuffers==25.1.24
googleapis-common-protos==1.66.0
grpcio==1.70.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
protobuf==5.29.3
requests==2.32.3
sympy==1.13.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
